Se arregló problma de parseo

+++ Entity/Persona.cs
@@ -3,7 +3,7 @@ namespace Entity {
    public class Persona {
       public string Identificacion { get; set; }
       public string Nombre { get; set; }
-      public int Edad { get; set; }
+      public decimal Edad { get; set; }
       public string Sexo { get; set; }
       public decimal Pulsacion { get; set; }
       public void CalcularPulsaciones ( ) {
